refactor spooling stuff into cluster_spooling
